// crates/desktop/src/generated/dto/SubAgentResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConfigSource } from "./ConfigSource";

export type SubAgentResponse = { name: string, description: string | null, instruction: string | null, source_path: string | null, source: ConfigSource | null, agent: string | null, };